+++ Patches/ChatCommandPatch.cs
@@ -1129,7 +1129,7 @@ internal class ChatCommands
             "窺視者" or "观察者" or "观察" => GetString("Observer"),
             "君主" => GetString("Monarch"),
             "預言家" or "预言家" or "预言" => GetString("Overseer"),
-            "驗屍官" or "验尸官" or "验尸" => GetString("Bloodhound"),
+            "驗屍官" or "验尸官" or "验尸" => GetString("Coroner"),
             "正義的追蹤者" or "正义追踪者" or "正义的追踪者" => GetString("Tracker"),
             "商人" => GetString("Merchant"),
             "總統" or "总统" => GetString("President"),
