+++ Resources/Lang/en_US.json
@@ -203,7 +203,7 @@
   "Observer": "Observer",
   "Monarch": "Monarch",
   "Overseer": "Overseer",
-  "Bloodhound": "Coroner",
+  "Coroner": "Coroner",
   "Tracker": "Tracker",
   "Merchant": "Merchant",
   "President": "President",
@@ -569,7 +569,7 @@
   "StealthInfo": "Killing Blinds Everyone in the Room",
   "PenguinInfo": "Drag your victims",
   "OverseerInfo": "Reveal roles of other players",
-  "BloodhoundInfo": "Find corpses and their killers",
+  "CoronerInfo": "Find corpses and their killers",
   "TrackerInfo": "Keep track of other players",
   "PresidentInfo": "You are in charge of the meeting",
   "MerchantInfo": "Sell add-ons and bribe killers",
@@ -866,7 +866,7 @@
   "MonarchInfoLong": "(Crewmates):\nAs the Monarch, you can knight players to give them an extra vote.\n\nYou cannot knight someone who already has extra votes.\n\nKnighted players appear with a golden name.\nIf a knighted player is alive, the Monarch cannot be guessed or exhiled.",
   "DovesOfNeaceInfoLong": "(Crewmates):\nWhen the Pacifist vents, they will reset the kill cooldown for every player with a kill button. When they become a Madmate, this ability will only work on Crewmates.",
   "OverseerInfoLong": "(Crewmates):\nAs a Overseer you have very limited vision but you can use your kill button to reveal the role of a nearby player. Use the kill button to start the reveal, a 「○」 will be displayed next to the reveal target. Stay near the target for a defined time to reveal his role, if you move too far away from the target the reveal will be aborted.",
-  "BloodhoundInfoLong": "(Crewmates):\nAs a Coroner you can't report corpses, instead after trying to report the corpse you will see an arrow leading you to the killer. If a meeting is called, the arrows disappear. Depending on the setting, the body you found cannot be reported.",
+  "CoronerInfoLong": "(Crewmates):\nAs a Coroner you can't report corpses, instead after trying to report the corpse you will see an arrow leading you to the killer. If a meeting is called, the arrows disappear. Depending on the setting, the body you found cannot be reported.",
   "TrackerInfoLong": "(Crewmates):\nAs a Tracker, you can vote for a player in the meeting, which will mark their position for you in the game with arrows. In addition, at the beginning of a meeting you will be shown in which room the player was last, if the option is activated.",
   "PresidentInfoLong": "(Crewmates):\nThe President has 2 abilities: End the meeting and Reveal identity.\n\n+ Ability 1: End the meeting - Type /finish in meetings as President to instantly end the meeting.\n+ Ability 2: Reveal identity - Type /reveal in meetings to reveal yourself. Revealing yourself will make it so every player can see that you are the President and you will become unguessable after typing the command. However, after the President has revealed themselves, whoever killed the President will have their kill CD greatly reduced on their next kill.",
   "MerchantInfoLong": "(Crewmates):\\As a merchant, you sell a random add-on to a random player for each task you complete. Each add-on sold earns you money. If you have a certain amount of money, you can avert the next killing attempt against you by bribing the murderer. The bribed player won't be able to kill you, but you don't know who it is. The bribe money used is lost and is not available for additional bribes.",
@@ -1842,9 +1842,9 @@
   "MediumshiperOnlyReceiveMsgFromCrew": "Receive messages only from <color=#8cffff>Crewmates</color> (including <color=#ff1919>Madmates</color> and <color=#cf6acd>Charmed</color> Players)",
   "DovesOfNeaceCooldown": "Ability Cooldown",
   "DovesOfNeaceMaxOfUseage": "Max Number of Ability Uses",
-  "BloodhoundArrowsPointingToDeadBody": "Arrows pointing to dead bodies",
-  "BloodhoundLeaveDeadBodyUnreportable": "Bodies the <color=#8B0000>Coroner</color> uses can't be reported",
-  "BloodhoundInformKillerBeingTracked": "Inform the Killer that he gets tracked",
+  "CoronerArrowsPointingToDeadBody": "Arrows pointing to dead bodies",
+  "CoronerLeaveDeadBodyUnreportable": "Bodies the <color=#8B0000>Coroner</color> uses can't be reported",
+  "CoronerInformKillerBeingTracked": "Inform the Killer that he gets tracked",
   "TrackerHideVote": "Hide <color=#3CB371>Tracker</color> Votes",
   "TrackerCanGetArrowColor": "Can See Colored Arrows",
 
@@ -2167,9 +2167,9 @@
   "DovesOfNeaceMaxUsage": "Ability use limit reached",
   "DovesOfNeaceSkillNotify": "Pacifist reset your kill cooldown",
   "BeRecruitedByJackal": "You have been recruited by the Jackal",
-  "BloodhoundTrackRecorded": "Track recorded",
-  "BloodhoundNoTrack": "Nothing to track",
-  "BloodhoundIsTrackingYou": "The Coroner is tracking you!",
+  "CoronerTrackRecorded": "Track recorded",
+  "CoronerNoTrack": "Nothing to track",
+  "CoronerIsTrackingYou": "The Coroner is tracking you!",
   "TrackerLastRoomMessage": "The evaluation of your tracking showed that the last room in which your target was located was:[{0}]",
   "MerchantAddonDelivered": "Add-on sold",
   "MerchantAddonSell": "The Merchant sold you a new Add-on",
