+++ Resources/roleColor.json
@@ -49,7 +49,7 @@
   "Observer": "#a8e0fa",
   "DovesOfNeace": "#007FFF",
   "Monarch": "#FFA500",
-  "Bloodhound": "#8B0000",
+  "Coroner": "#8B0000",
   "Tracker": "#3CB371",
   "Merchant": "#D27D2D",
   "Retributionist": "#228B22",
